+++ package/tests/test_PartSegCore/test_segmentation_proceses.py
@@ -9,7 +9,7 @@ from PartSegCore.algorithm_describe_base import ROIExtractionProfile
 from PartSegCore.analysis.algorithm_description import analysis_algorithm_dict
 from PartSegCore.analysis.load_functions import UpdateLoadedMetadataAnalysis
 from PartSegCore.json_hooks import check_loaded_dict
-from PartSegCore.segmentation.algorithm_base import SegmentationAlgorithm
+from PartSegCore.segmentation.algorithm_base import ROIExtractionAlgorithm
 from PartSegImage import TiffImageReader
 
 
@@ -34,7 +34,7 @@ class TestSegmentation:
 
         val: ROIExtractionProfile
         for val in data.values():
-            algorithm: SegmentationAlgorithm = analysis_algorithm_dict[val.algorithm]()
+            algorithm: ROIExtractionAlgorithm = analysis_algorithm_dict[val.algorithm]()
             algorithm.set_image(image)
             algorithm.set_mask(image.mask.squeeze())
             algorithm.set_parameters(**val.values)